Adde projected gradient descent python file
Currently it takes as input a list of goal (Vector 3r) classes and iterates through that, feel free to change that if you want something else. Does a quadratic gradient descent for far away goals with some learning rate eta, if the goal is close does a conic gradient descent (makes it go slower).